Extract mock cfe result logic to mixin
Is used by a couple of controllers now
so we should ensure the same behaviour.
